requirements.txt
Adafruit-GPIO==1.0.4
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
aggdraw==1.3.16
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
attrs==23.1.0
automationhat==0.2.0
beautifulsoup4==4.9.3
bing-image-downloader==1.1.2
blender==1.4
blinker==1.4
blinkt==0.1.2
broadlink==0.18.3
buildhat==0.5.12
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
contourpy==1.1.1
cryptography==3.3.2
cupshelpers==1.0
cycler==0.12.1
Cython==0.29.27
darkdetect==0.6.0
dbus-python==1.2.16
distro==1.5.0
docopt==0.6.2
docutils==0.16
drumhat==0.1.0
editpi==1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.1.2
fonttools==4.43.1
fourletterphat==0.1.0
gpiozero==1.6.2
html5lib==1.1
idna==2.10
imageio==2.31.4
importlib-resources==6.1.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.3.2
kiwisolver==1.4.5
lazy-object-proxy==0.0.0
lazy_loader==0.3
logilab-common==1.8.1
lxml==4.6.3
markdown-it-py==2.1.0
MarkupSafe==1.1.1
matplotlib==3.8.0
mccabe==0.6.1
mcpi==1.2.1
mcpit==1.1
mdurl==0.1.2
mfrc522==0.0.7
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mutf8==1.0.6
mypy==0.812
mypy-extensions==0.4.3
networkx==3.1
numpy==1.26.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
packaging==23.1
Panda3D==1.10.13
panda3d-gltf==1.0.0
panda3d-simplepbr==0.10
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==10.0.1
psd-tools==1.9.28
psutil==5.8.0
pyamiibo==0.2
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
pygame==1.9.6
Pygments==2.14.0
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyNBT==3.1.0
pyOpenSSL==20.0.1
pyparsing==3.1.1
pyperclip==1.8.2
pypresence==4.2.1
pyqtdarktheme==1.1.0
pyserial==3.5
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
PyWavelets==1.4.1
qtwidgets==0.18
rainbowhat==0.1.0
raspberrypi-python-tm1637==1.3.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
rich==13.3.1
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-emu==1.2
sense-hat==2.4.0
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
SPI-Py==1.0
spidev==3.5
ssh-import-id==5.10
termcolor==2.2.0
thonny==4.0.1
threadpoolctl==3.2.0
tifffile==2023.9.18
toml==0.10.1
touchphat==0.0.1
tqdm==4.62.3
twython==3.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
unicornhathd==0.0.4
urllib3==1.26.5
ursina==5.2.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xlib==0.21
zipp==3.17.0
